+++ TDEngine2/plugins/OGLGraphicsContext/include/COGLResources.h
@@ -274,6 +274,9 @@ namespace TDEngine2
 			E_RESULT_CODE Init(IGraphicsContext* pGraphicsContext, const TInitTextureImplParams& params) override;
 
 			E_RESULT_CODE Resize(U32 width, U32 height, U32 depth = 1) override;
+			E_RESULT_CODE Transition(E_RESOURCE_LAYOUT newLayout) override;
+
+			E_RESULT_CODE SetHandle(TTextureHandleId handle, const CPassKey<CBaseGraphicsObjectManager>& passkey) override;
 
 			GLuint GetTextureHandle();
 
@@ -289,8 +292,8 @@ namespace TDEngine2
 			E_RESULT_CODE _onFreeInternal() override;
 		protected:
 			TInitTextureImplParams     mInitParams;
-
-			GLuint mTextureHandle = 0;
+			TTextureHandleId           mHandle;
+			GLuint                     mTextureHandle = 0;
 	};